rename module to sensors
