+++ metadrive/component/sensors/base_camera.py
