+++ metadrive/component/sensors/distance_detector.py
