+++ metadrive/component/sensors/vehicle_panel.py
