+++ metadrive/policy/idm_policy.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from metadrive.component.lane.point_lane import PointLane
-from metadrive.component.vehicle_module.PID_controller import PIDController
+from metadrive.component.vehicle.PID_controller import PIDController
 from metadrive.policy.base_policy import BasePolicy
 from metadrive.policy.manual_control_policy import ManualControlPolicy
 from metadrive.utils.math import not_zero, wrap_to_pi, norm
