+++ metadrive/policy/lange_change_policy.py
@@ -1,6 +1,6 @@
 import gymnasium as gym
 
-from metadrive.component.vehicle_module.PID_controller import PIDController
+from metadrive.component.vehicle.PID_controller import PIDController
 from metadrive.engine.engine_utils import get_global_config
 from metadrive.policy.env_input_policy import EnvInputPolicy
 from metadrive.utils.math import wrap_to_pi
