+++ metadrive/tests/scripts/capture_obs.py
@@ -1,6 +1,6 @@
 from metadrive.component.map.base_map import BaseMap
 from metadrive.component.map.pg_map import MapGenerateMethod
-from metadrive.component.vehicle_module.depth_camera import DepthCamera
+from metadrive.component.sensors.depth_camera import DepthCamera
 from metadrive.envs.metadrive_env import MetaDriveEnv
 from metadrive.utils import setup_logger
