+++ metadrive/tests/test_functionality/test_navigation.py
@@ -1,6 +1,6 @@
 from metadrive.component.map.base_map import BaseMap
 from metadrive.component.map.pg_map import MapGenerateMethod
-from metadrive.component.vehicle_module.PID_controller import PIDController
+from metadrive.component.vehicle.PID_controller import PIDController
 from metadrive.envs.metadrive_env import MetaDriveEnv
