+++ metadrive/tests/tools/adjust_collision_model.py
@@ -1,8 +1,8 @@
 import numpy as np
 from metadrive.component.vehicle.base_vehicle import BaseVehicle
-from metadrive.component.vehicle_module.mini_map import MiniMap
-from metadrive.component.vehicle_module.rgb_camera import RGBCamera
-from metadrive.component.vehicle_module.vehicle_panel import VehiclePanel
+from metadrive.component.sensors.mini_map import MiniMap
+from metadrive.component.sensors.rgb_camera import RGBCamera
+from metadrive.component.sensors.vehicle_panel import VehiclePanel
 
 from metadrive.envs.metadrive_env import MetaDriveEnv
 from metadrive.utils import setup_logger
